+++ gqcp/include/QCModel/HF/Stability/StabilityProperties/CMakeLists.txt
@@ -1,5 +1,5 @@
 target_sources(gqcp
     PRIVATE
         GHFStabilityProperties.hpp
-        StabilityProperties.hpp
+        Stability.hpp
 )
